add config option for synthetic impressions
